Split chat into component
